+++ docs/.vitepress/vitepress/components/doc-content/vp-table-of-content.vue
@@ -43,13 +43,13 @@ const sponsor = computed(() => sponsorLocale[lang.value])
         class="mt-8 toc-ads flex flex-col"
         item-style="width: 180px; height: 55px;"
       /> -->
-      <p class="text-14px font-300 color-$text-color-secondary">
-        {{ sponsor.sponsoredBy }}
-      </p>
-      <sponsors-button class="sponsors-button mt-4 w-100%" />
-      <sponsor-right-big-logo-list />
-      <sponsor-right-logo-small-list />
-      <sponsor-right-text-list />
+      <!--      <p class="text-14px font-300 color-$text-color-secondary">-->
+      <!--        {{ sponsor.sponsoredBy }}-->
+      <!--      </p>-->
+      <!--      <sponsors-button class="sponsors-button mt-4 w-100%" />-->
+      <!--      <sponsor-right-big-logo-list />-->
+      <!--      <sponsor-right-logo-small-list />-->
+      <!--      <sponsor-right-text-list />-->
     </nav>
   </aside>
 </template>