+++ docs/.vitepress/vitepress/components/vp-content.vue
@@ -49,6 +49,6 @@ watch(
       <template #content-top><slot name="content-top" /></template>
       <template #content-bottom><slot name="content-bottom" /></template>
     </VPDocContent>
-    <VPFooter v-if="!isHeroPost" />
+    <!--    <VPFooter v-if="!isHeroPost" />-->
   </main>
 </template>