+++ docs/examples/next-card/shadow.vue
@@ -1,7 +1,0 @@
-<template>
-  <div class="flex flex-wrap gap-4">
-    <el-card style="width: 480px" shadow="always">Always</el-card>
-    <el-card style="width: 480px" shadow="hover">Hover</el-card>
-    <el-card style="width: 480px" shadow="never">Never</el-card>
-  </div>
-</template>