+++ docs/examples/next-card/with-icon-title-desc-vertical.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <ElNextCard
+      header-bordered
+      type="vertical"
+      title="列表标题"
+      desc="列表内容的描述性文字"
+      style="max-width: 300px"
+    >
+      <template #avatar>
+        <img src="https://tdesign.gtimg.com/site/avatar-boy.jpg" />
+      </template>
+    </ElNextCard>
+  </div>
+</template>